boxFilter implementation moved to separate file
